refactor: move telescope config into module
