+++ lua/jeremiah/init.lua
@@ -2,6 +2,7 @@ require("jeremiah.util")
 require("jeremiah.packer")
 require("jeremiah.colors")
 require("jeremiah.ccc")
+require("jeremiah.telescope")
 require("jeremiah.ripgrep")
 require("jeremiah.globals")
 require("jeremiah.treesitter")
